Added Task saving and loading
Tasks implement saving and loading functions. A result is saved to a specified
Storage location, and then loaded instead of recomputed.

To replace Task.run for a loading functions when recomputing the same graph,
the switch must happen on the dask optimize step, not on graph building (ie, Task.__new__).

Storage define a key-value store, which must implement an exist method to check
for the existence of a given key (to replace the Task.run with a loading method).

Serializer classes implement dumps and loads functions to transform a Task
result before saving to Storage.