Added save, encode and decode methods to Task
When Task.save is True, the result of its computation is saved
(when a Storage class has been created). The result is previously
encoded with Task.encode, which performs a serialization, compression,
and encryption of that value. When loading that result from storage,
the inverse process is done by Task.decode.

By default, serialization is done with cloudpickle (which is not suitable
for long term storage), compression with zstandard, and no encryption is used.
The encoding process can be adapted by setting Task.serializer, Task.compressor,
and Task.encrypter to modules or classes that implement the appropiate functions,
or by overriding the encode and decode methods.