Caching delayed run construction
Profiling Task instantiation, 75% of the time is spent on `func = delayed(self, ...)`.

Instead of creating a new Delayed object every time we create a Task instance,
now we create a single Delayed on class construction.

Since we later need access to the instance, this cached Delayed takes "self" as a parameter.

But since Dask considers callables as "tasks", we need to remove the Task.__call__ method.

Additionally, to prevent Dask from traversing the instance, we wrap it in a `literal`.

This reduced graph construction a factor x6.