Pass BoundArguments to ArgumentHasher
Removes ambiguities for positional_or_keyword parameters,
which produced different hashes if an argument was
passed by position or by keyword.

Having to compute BoundArguments from the signature,
we can error earlier on missing or unexpected arguments.

Simplifies excluding parameters from hash.

But it is x10 slower, which might have an impact on large graphs.